feat - Crud do calendario

+++ projeto-tfg/src/main/kotlin/school/sptech/projetotfg/repository/AtividadeRepository.kt
@@ -3,5 +3,5 @@ package school.sptech.projetotfg.repository
 import org.springframework.data.jpa.repository.JpaRepository
 import school.sptech.projetotfg.domain.atividades.Atividade
 
-interface AtividadeRepository:JpaRepository<Atividade, Int> {
+interface AtividadeRepository:JpaRepository<Atividade, Long> {
 }
